# docs/swagger-v2.yml
openapi: 3.0.3
info:
  title: Hairsap Api
  description: |-
    **This is the documentation for hairsap api.**

    NOTES
    - Amounts should be sent as `integers` not `floats`, so $100.95 would be held as `10095` in any transaction amount or balance fields.
    - Dates should be sent in ISO format e.g `2022-09-09T07:52:39.407Z`

  # termsOfService: http://swagger.io/terms/
  contact:
    name: the Developer
    # url:
    email: jideadedejifirst@gmail.com
  version: 1.0.0
...
servers:
  - url: http://dev.hairsap.com:3000
  - url: http://localhost:4000
...
paths:
...
  /admin/pros:
    post:
      tags:
        - Admin
      summary: pros
      description: pros
      operationId: pros
      parameters: []
      requestBody:
        description: Pros
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminProsRequest'
        required: true
      responses:
        '200':
          description: Status 200 Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminProsResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
      security:
        - BearerAuth: []
  /admin/users:
    post:
      tags:
        - Admin
      summary: users
      description: users
      operationId: users
      parameters: []
      requestBody:
        description: users
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUsersRequest'
        required: true
      responses:
        '200':
          description: Status 200 Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUsersResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
      security:
        - BearerAuth: []
  /admin/users/{userId}:
    get:
      tags:
        - Admin
      summary: users id
      description: users id
      operationId: usersid
      parameters: 
        - name: userId
          in: path
          required: true
          allowEmptyValue: false
          schema:
            type: number
            example: 1
      responses:
        '200':
          description: Status 200 Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUsersIdResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
      security:
        - BearerAuth: []
  /admin/pros/{userId}:
    get:
      tags:
        - Admin
      summary: pros user id
      description: pros user id
      operationId: prosuserid
      parameters: 
        - name: userId
          in: path
          required: true
          allowEmptyValue: false
          schema:
            type: number
            example: 1
      responses:
        '200':
          description: Status 200 Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminProsUserIdResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
      security:
        - BearerAuth: []
...
  /users:
    patch:
      tags:
        - User
      summary: user patch 
      description: user patch 
      operationId: userpatch 
      parameters: []
      requestBody:
        description: user patch
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPatchRequest'
        required: true
      responses:
        '200':
          description: Status 200 Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPatchResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
      security:
        - BearerAuth: []
  /users/faceid:
    post:
      tags:
        - User
      summary: user faceid 
      description: user faceid 
      operationId: faceid 
      parameters: []
      requestBody:
        description: user faceid
        content:
          mutltipart/form-data:
            schema:
              $ref: '#/components/schemas/UserfaceIdRequest'
        required: true
      responses:
        '200':
          description: Status 200 Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserfaceIdResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
      security:
        - BearerAuth: []
  /profilephoto:
    post:
      tags:
        - User
      summary: profile photo  
      description: profile photo
      operationId: profilephoto 
      parameters: []
      requestBody:
        description: profile photo
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfilePhotoRequest'
        required: true
      responses:
        '200':
          description: Status 200 Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfilePhotoResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
      security:
        - BearerAuth: []
  /subscribe:
    post:
      tags:
        - User
      summary: subscribe  
      description: subscribe
      operationId: subscribe 
      parameters: []
      requestBody:
        description: subscribe
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserSubscribeRequest'
        required: true
      responses:
        '200':
          description: Status 200 Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSubscribeResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
      security:
        - BearerAuth: []
  /subscriptions:
    post:
      tags:
        - User
      summary: subscriptions  
      description: subscriptions
      operationId: subscriptions 
      parameters: []
      requestBody:
        description: subscriptions
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserSubscriptionsRequest'
        required: true
      responses:
        '200':
          description: Status 200 Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSubscriptionsResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
      security:
        - BearerAuth: []
  /me:
    get:
      tags:
        - User
      summary: me  
      description: me
      operationId: me 
      parameters: []
      responses:
        '200':
          description: Status 200 Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserMeResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
      security:
        - BearerAuth: []
  /card:
    get:
      tags:
        - User
      summary: card  
      description: card
      operationId: card 
      parameters: []
      responses:
        '200':
          description: Status 200 Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserCardResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
      security:
        - BearerAuth: []
  /webhook/paystack:
    post:
      tags:
        - Others
      summary: Webhook Paystack Integration
      description: Webhook Paystack Integration
      operationId: webhookpaystack
      parameters: []
      requestBody:
        description: Webhook Paystack Integration
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookPaystackRequest'
        required: true
      responses:
        '200':
          description: Status 200 Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookPaystackResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
      security:
        - BearerAuth: []
  /verify_transaction:
    get:
      tags:
        - Others
      summary: verify transaction
      description: verify transaction
      operationId: verifytransaction
      parameters: []
      
      responses:
        '200':
          description: Status 200 Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/verifyTransactionResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
      security:
        - BearerAuth: []
  /services:
    get:
      tags:
        - Others
      summary: Get all services and sub services
      description: Get all services and sub services
      operationId: services
      parameters: []
      responses:
        '200':
          description: Status 200 Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServicesResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
      security:
        - BearerAuth: []
components:
  schemas:
...
    ForbiddenError:
      required:
        - message
      properties:
        message:
          type: string
          default: Forbidden
    NotFoundError:
      required:
        - message
      properties:
        message:
          type: string
          default: Not Found
    UnauthorizedError:
      required:
        - message
      properties:
        message:
          type: string
          default: Unauthorized
...
    Meta: 
     type: object
     properties:
        total:
          type: number
        skipped:
          type: number
        perPage:
          type: number
        page:
          type: number
        pageCount:
          type: number
...
    AdminProsRequest:
      type: object
      required:
        - name
        - serviceId
        - page
        - perPage
      properties:
        name:
          type: string
          description: name
          example: 'pro'
        serviceId:
          type: number
          description: serviceId
          example: 1
        page:
          type: number
          description: page
          example: 1
        perPage:
          type: number
          description: perPage
          example: 10
    AdminProsResponse:
      type: object
      properties:
        meta: 
          $ref: '#/components/schemas/Meta'
        data:
          type: array
          items:
            type: object
            properties:
              userId:
                type: number
              businessName:
                type: string 
              name:
                type: string 
              email:
                type: string    
              deactivated:
                type: boolean  
              createdAt:
                type: string  
              profilePhotoUrl:
                type: string  
    AdminUsersRequest:
      type: object
      required:
        - name
        - page
        - perPage
      properties:
        name:
          type: string
          description: name
          example: 'james'
        page:
          type: number
          description: page
          example: 1
        perPage:
          type: number
          description: perPage
          example: 10
    AdminUsersResponse:
      type: object
      required:
        - name
        - userId
        - createdAt
      properties:
        name:
          type: string
          description: name
          example: 'pro'
        address:
          type: string
          description: address
          example: 'No `, Ikeja Street, Lagos'
        userId:
          type: number
          description: id of user
          example: 1
        createdAt:
          type: string
          example: '2022-08-26T15:28:13.528Z'
        email:
          type: string
        phone:
          type: string
        profilePhotoUrl:
          type: string
        verified:
          type: boolean
    AdminUsersIdResponse:
      type: object
      required:
        - userId
      properties:
        amountSpent:
          type: number
        averageRatings:
          type: number
        subscriptions:
          type: number
        totalBookings:
          type: number
        user:
          type: object
          properties:
            userId:
              type: number
              description: userId
              example: 1
            name:
              type: string
            profilePhotoUrl:
              type: string
            email:
              type: string
            createdAt:
              type: string
            verified:
              type: boolean
            phone:
              type: string
            address:
              type: string
    AdminProsUserIdResponse:
      type: object
      properties:
        dailyBookingCount:
          type: number
        dailyBookingSum:
          type: number
        weeklyBookingCount:
          type: number
        weeklyBookingSum:
          type: number
        monthlyBookingCount:
          type: number
        monthlyBookingSum:
          type: number
        allBookingCount:
          type: number
        allBookingSum:
          type: number
        subscriptions:
          type: number
        averageRatings:
          type: number
        ratings:
          type: number
        taskTarget:
          type: object
          properties:
            day:
              type: number
            week:
              type: number
            month:
              type: number
        earnings:
          type: object
          properties:
            day:
              type: number
            week:
              type: number
            month:
              type: number
        user:
          type: object
          properties:
            userId:
              type: number
              description: userId
              example: 1
            name:
              type: string
            profilePhotoUrl:
              type: string
            email:
              type: string
            createdAt:
              type: string
            deactivated:
              type: boolean
            phone:
              type: string
            address:
              type: string
            service:
              type: string
            account:
              type: object
              properties:
                accountId:
                  type: number
                accountNumber:
                  type: string
                accountName:
                  type: string
                bankName:
                  type: string
        latestBookings:
          type: array
          items:
            type: object
            properties:
              bookingId:
                type: number
              createdAt:
                type: string
              bookedSubServices:
                type: array
                items:
                  type: object
                  properties:
                    subServiceId:
                      type: number
                    serviceId:
                      type: number
                    name:
                      type: string
                    photoUrl:
                      type: string
                    price:
                      type: number
                    createdAt:
                      type: string
...
    UserPatchRequest:
      type: object
      properties:
        address:
          type: string
    UserPatchResponse:
      type: object
    UserfaceIdRequest:
      type: object
      required:
        - userId
        - faceId
        - proId
      properties:
        userId:
          type: number
          description: userId
          example: 1
        proId:
          type: number
          description: proId
          example: 1
        file:
            type: array
            items:
              type: string
              format: binary
        faceId:
          type: string
          format: binary #base64
    UserfaceIdResponse:
      type: object
      required:
        - userId
        - role
        - proId
        - faceIdPhotoKey
        - faceIdPhotoOriginalFileName
      properties:
        userId:
          type: number
          description: userId
          example: 1
        role:
          type: string
          description: role
          example: 'user'
        proId:
          type: number
          description: proId
          example: 1
        faceIdPhotoKey:
          type: string
          description: faceIdPhotoKey
          example: 'photo key'
        faceIdPhotoOriginalFileName:
          type: number
          description: faceIdPhotoOriginalFileName
          example: 'Photo Original File Name'
    UserProfilePhotoRequest:
      type: object
      required:
        - userId
        - proId
        - profilePhotoKey
        - profilePhotoOriginalFileName
        - profilePhotoUrl
      properties:
        userId:
          type: number
          description: userId
          example: 1
        proId:
          type: number
          description: proId
          example: 1
        profilePhotoKey:
          type: string
          description: profilePhotoKey
          example: 'profile Photo Key'
        profilePhotoOriginalFileName:
          type: number
          description: profilePhotoOriginalFileName
          example: 'Profile Photo Original FileName'
        profilePhotoUrl:
          type: number
          description: profilePhotoUrl
          example: 'profile Photo Url'
    UserProfilePhotoResponse:
      type: object
      required:
        - userId
        - profilePhotoKey
        - faceIdPhotoKey
        - profilePhotoUrl
      properties:
        userId:
          type: number
          description: userId
          example: 1
        profilePhotoKey:
          type: string
          description: profilePhotoKey
          example: 'profilePhotoKey'
        faceIdPhotoKey:
          type: string
          description: faceIdPhotoKey
          example: 'photo key'
        profilePhotoOriginalFileName:
          type: number
          description: profilePhotoOriginalFileName
          example: 'Photo Original File Name'
        profilePhotoUrl:
          type: number
          description: profilePhotoUrl
          example: 'Profile Photo Url'
    UserSubscribeRequest:
      type: object
      required:
        - userId
        - proId
      properties:
        userId:
          type: number
          description: userId
          example: 1
        proId:
          type: number
          description: proId
          example: 1
    UserSubscribeResponse:
      type: object
      required:
        - userId
        - proId
      properties:
        userId:
          type: number
          description: userId
          example: 1
        proId:
          type: number
          description: proId
          example: 1
    UserSubscriptionsRequest:
      type: object
      required:
        - userId
      properties:
        userId:
          type: number
          description: userId
          example: 1
    UserSubscriptionsResponse:
      type: object
      required:
        - userId
      properties:
        userId:
          type: number
          description: userId
          example: 1
    UserMeResponse:
      type: object
      required:
        - userId
      properties:
        userId:
          type: number
          description: userId
          example: 1
    UserCardResponse:
      type: object
      required:
        - userId
      properties:
        userId:
          type: number
          description: userId
          example: 1
    WebhookPaystackRequest:
      type: object
      required:
        - userId
      properties:
        userId:
          type: number
          description: userId
          example: 1
    WebhookPaystackResponse:
      type: object
      required:
        - userId
        - email
        - event
        - reason
      properties:
        userId:
          type: number
          description: userId
          example: 1
        email:
          type: string
          description: email
          example: 'email'
        event:
          type: string
          description: event
          example: 'event'
        reason:
          type: string
          description: reason
          example: 'reason'
    verifyTransactionResponse:
      type: object
      required:
        - userId
      properties:
        userId:
          type: number
          description: userId
          example: 1
...
  responses:
    InternalError:
      description: Status 500 Response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/InternalError'
    ValidationError:
      description: Status 400 Response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
    ForbiddenError:
      description: Status 403 Response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ForbiddenError'
    NotFoundError:
      description: Status 404 Response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/NotFoundError'
    UnauthorizedError:
      description: Status 401 Response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UnauthorizedError'
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer